sol-pin-mux: Add pin multiplexer capabilities
This patch adds support to pin multiplexer instructions that can be used
by nodes of specific platforms (galileo, edison and others) to properly
set a given pin to work in a given mode.

Multiplexer instructions can be loaded in runtime (if support for is
available) and will usually happen during the specific module initialization.
This way Soletta don't need either host code or be modified to support
additional platforms that need this kind of setup to work properly.

This patch also add the instructions of Galileo Gen2 multiplexer.