platform-linux: move fork_run to linux-common.
The fork_run implementation was written in a linux-generic way,
although only used by linux-micro modules and thus it lived in such
platform.

As we plan to have a "subprocess" node type, it would be nice to have
it using the same infrastructure, then it need to be moved to a
linux-common file.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>