flow: add process/stdin node type.
The 'process' module will offer types:
    * stdin: reads chunks (blobs)
    * stdout/stderr: writes chunks (blobs)
    * subprocess: equivalent to popen().

The first type is 'stdin' and it's a bit more complex than one would
anticipate as we may have multiple nodes and ports listening to stdin
and we need to use a central reader to allow that.

The central reader is based on chunks (blobs) that actual reading is
limited in time and memory, once those are exhausted or the read would
block, then the chunk is finalized and sent on 'OUT' ports of
interested nodes.

There is a 'CLOSED' port to report when stdin is closed, there is an
initial packet to say what's the current status at the time of the
connection.

This module will conflict with 'keyboard', if both are open, they will
all read from the same file descriptor and cause issues. I'll try to
solve this in a future commit by using dup() to a new fd and closing
stdin(0) so the second to open will fail. Right now let's live with
that -- it will also be an issue for stdout/stderr and console.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>