flow:process: stdout/stderr implementation
These two nodes are part of the process node. They have an input
port (blob) where they receive the that will be written. As they
use nonblocking file descriptors all data comes is put in queue
and used as soon the file descriptor is able to be written.

There is a 'CLOSED' port to report when stdout/stderr is closed, there
is an initial packet to say what's the current status at the time of the
connection.

The only difference between the nodes is the file descriptor they work.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>