thirdparty: add duktape v1.2.1
Duktape[1] is an embeddable Javascript engine, with focus on
portabillity and compact footprint.

As suggested by the guide, duktape instalation is done by
adding C source and header to the build.

We're adding this in the thirdparty directory, in order to use
it in the next commit, where soletta meets js node.

[1] - http://duktape.org / https://github.com/svaarala/duktape

Signed-off-by: Ricardo de Almeida Gonzaga <ricardo.gonzaga@intel.com>