flow:process: subprocess node
This node allows to spawn aprocesses and redirect their
input/output/error to its ports. It takes two parameters in the
options, the 'command' is the program (including its parameters)
that will be spawned and the 'start' (optional) that tells to the
node if the process should be started at the node's open or
only after receive a packet in its 'START' port.

There are seven ports:

'START'  (in any)   - Start the child (if it's not running)
'STOP'   (in any)   - Send a SIGTERM to the child
'STDIN'  (in blob)  - Write the data into child's stdin
'SIGNAL' (in int)   - Send the specified signal to the child
'STDOUT' (out blob) - The data written by the child process in
                      its stdout
'STDERR' (out blob) - The data written by the child process in
                      its stderr
'PID'    (out int)  - The child's pid

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>