js: add a way to create node types using JS
Allow the usage of Javascript language to create new and
customizable node types.

A JS node type is specified with one object containing each
input and output port declarations (name and type) and its
callback functions that will be trigged on the occurrence
of certain events like input/output ports processes, open/close
processes, so forth and so on.

Signed-off-by: Ricardo de Almeida Gonzaga <ricardo.gonzaga@intel.com>