flow: Add array ports
Array ports are declared on the JSON file as NAME[size], where size has
to be a positive integer. This will generate a single port description
and type, but for the framework, one individual port for each index
exists.

Connections to and from array ports have to be explicit. Trying to
connect to an array port without specifying an index will result in an
error. If this is too much of a hindrance, we can look for another
solution later, but for now it makes things much simpler.

Export a port an array port can be done by just using the name, in
which case the whole array is exported, or by exporting individual
indices, which exports them as single ports.

Helper functions were added to get a port description from the port
index as given by the framework, since these will not longer necessary
match to the position in the array.