flow: Convert some node types to use array ports
The node types for byte-to-bits, bits-to-byte, int-compose and
int-decompose now use array ports.

Functionality doesn't change, but now the definitions of the port
indices are used to get the internal index to be used with internal
arrays or to get the proper output port. This should be more robust in
case of future changes in the node type, ensuring that the indices used
to access internal memory remain sane if the base port numbers change.