kconfig
This patch introduces the kconfig as soletta build system configurator.

The work flow is pretty similar to linux kernel's one, make menuconfig,
make oldconfig, make allyesconfig, make allmodconfig and so on.

The object and modules configuration are very similar too. The main files
to look at while adding new modules and configuration options are the
Kconfig and Makefile's spread across the source tree, where in the
Kconfig files one will add the module vs. objects configuration i.e:

obj-$(CONFIG_MY_OPTION) += my-module.mod
obj-my-module-$(CONFIG_MY_OPTION) += my-obj.o my-obj2.o

Pay attention to the my-module.mod "aberration", this .mod suffix is
internally used to indicate this is a "soletta object", I have used
this notation to avoid confusion between shared and builtin modules,
while configuring the module may be either shared or static, so
it's simply a "soletta object" beforehand.

The second file class(Kconfig) is pretty the same as in the linux kernel
source code. The main difference is the "object types", it may be one of:
bin-*-y, obj-*-[y,m], test-*-y. The first one is used to build binaries
like src/bin/, the second one defines a module and the third are reserved
for unit tests.

We also support external(out of tree) modules, and the way to build it
is exactly the same as in the linux kernel, write you Makefile(in the
external dir) and do the following:

make -C $PATH_TO_SOLETTA_SRC M=$PWD

To install soletta we have a target "install" and an special variable
DESTDIR, this variable works the same way as in the autotools world.
Having said that, a typical installation invocation would look like:

make install DESTDIR=~/local-install/

If DESTDIR's not provided the build system will try to install in the
sysroot.

The script autoconf.py is the autotools's autoconf replacement, it finds
the libs dependencies and properly  set the HAVE_* variables and the CFLAGS.
These variables are defined in two generated files Kconfig.gen and
Makefile.gen.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>