sol-flow-node-type-gen.py is full of help and uses default schema.
Generate the final sol-flow-node-type-gen.py replacing the pkgdatadir
variable so we know where to look for the schema, removing the burden
to specify this for external users.

Rename the too-generic 'input' term to 'genspec', so it's clear it is
the generator specification we're talking about. This will help in a
future commit where this tool will also generate the JSON
specification that is used by sol-fbp-generator,
sol-flow-node-type-find.py and other tools (currently generated from
the binary .so modules).

And take the time to document the arguments and the tool.