flow: modify how it deals with initial packets
Initial packets were supported as a special case for connect()
methods of node types. They were a special case because it was
possible to send different packets for each destination node.

But it causes a big trouble since it interrupt the sequence
of calls of connect() methods calling process() methods
between them. When a node type was sending a packet on connect
it was dispatching to connected node imediatelly, calling its
process() method.

It was breaking the expected behaviour, affecting some node types
already, like true/false counter.

Instead of that, to send initial packets now, it's enough
to use regular API to send packets at open() node types methods().

After all connections are done, this initial packets are going to
be dispatched to all the connected nodes. So they won't be
a special case anymore, won't break expected order of methods, and
would be good enough for all our cases so far.