sol-flow-node-type-gen.py now generates the JSON definitions.
Generating the JSON definitions from the actual binary is cumbersome
and won't work for some cross compilations, particularly for different
OSes.

Then we simplify and generate the JSON from the input genspec as it's
a simple translation.

This should also speed up the build as we execute one less binary.