flow: Drop REPLACE_PACKET output port flag
This flag was added as a way for out ports to replace the last sent packet
It was used to indicate that they replace the last sent
packet when more than one sol_flow_node_packet_send() call
happens during the same process cycle.

But in many cases receiving the all values was expected.
It had a meaning involved. As many packets are used to feed input
ports that receive "any" packet type, the simple pulse is information.

And lots of workaround emerged. So it's better to always send
packets and provide node types that could filter values.
So all the use cases can be covered.