modules: remove checks to avoid repeated values
For the same reason we've drop REPLACE_FLAG, we're now removing
these checks to avoid repeated packets.
They may have use depending on the flow.
For cases where they should be avoided, we'll provide
a family of node types to filter repeated packets, so
all the use cases can be covered.