Initial port of Contiki
For now just mainloop timeout and idle is working, everything else is just
stubs that are necessary to compile.

There is a mainloop-impl-common that is the mainloop code that was copy from
Riot and Posix implemenations, in a further patch the duplicated code will be
removed and replaced by a call to the mainloop-impl-common version.