util/resolver-conffile: guess install root dir
This patch introduces an util API to guess the current installation
root dir, it's basically using the new - comming - internal build
structure where we have a mirrored installation directory - by
default within $(top_srcdir)/build/. If the current process's
"progname" is has a soletta_sysroot we know we're running from
inside the build dir.

This API's first user is the sol-flow-resolver-conffile module, it
uses the guessed rootdir to compose the flow module path.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>