calamari: Get rid of hardcoded GPIO pins
Do the evolution, baby!
Calamari 7seg and RGB led nodes were the first container nodes we
created. As the infrastructure for this kind of nodes improved during
the time, we managed to remove the number of hardcoded things, now
it's time to remove the GPIO pin numbers, making use of the
child_opts_set hook.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@intel.com>