flow: properly parse rgb options if declared per key
It was working fine if rgb option was declared as
"X|Y|Z", but if it was declared as "red:X|green:Y|blue:Z"
it was ending up with default values.

It happens because field_cnt shouldn't start from 0 in switch,
since if no value was declared, it doesn't make
sense to declare the option. Also, if values were
assigned by ASSIGN_KEY_VAL field_cnt would stay 0, and
the whole option would be set to default values

Also, add a test for it.