flow-parser: Delete builders from the top
When an FBP file makes use of others through the DECLARE keyword, those
others will be parsed and added to the builders vector in the order
they are found, before the current file is finished parsing and added
to the same vector. This means that parent builders will end up in the
vector after their children, with the top-level one at the end.

When nodes are opened, their parents own the reference to their options
pointers, and those pointers will be freed as the nodes they belong to
are deleted, which will happen automatically when the parent is
deleted.

Since freeing of these options may end up referencing their respective
node-types, those types need to be alive at the time the options are
freed, and so we need to make sure the builders are not deleted before
their parents.

Signed-off-by: Iván Briano <ivan.briano@intel.com>