builder: Support exporting a sub-node's option
Export sub-node options as options from the parent, so that a type
created with the flow builder can be configurable.

This will create the options description based on the descriptions of
the respective sub-nodes, so if the option exported is marked as
required, the parent's will be too, making the type itself require the
passing of valid options.
Accordingly, if the option is not required, the default value from the
sub-node will be copied to the parent's.

When the node is instantiated, all its children will be opened with a
copy of the options the type was created with, modified with the values
of the exported options as received by the parent.

Signed-off-by: Iván Briano <ivan.briano@intel.com>