builder: Support exporting a sub-node's option
Export sub-node options as options from the parent, so that a type
created with the flow builder can be configurable.

This will create the options description based on the descriptions of
the respective sub-nodes, and the default value will be set from the
value of the option the sub-node is created with. If that specific
instance of the node has no options, the default from the type will be
used.

When the node is instantiated, all its children will be opened with a
copy of the options the type was created with, modified with the values
of the exported options as received by the parent.

Signed-off-by: Iván Briano <ivan.briano@intel.com>