log: embed log domain inits on node generation scripts
This way freeing the user to have to worry with log domains for their
nodes. If a node is built-in, the log domain is the global one. If not,
a domain following the node's name-space is automatically created from
the .json declaration, with the declaration on its *-gen.h. The
difference now is that this file has to be the 1st Soletta header
getting an #include in a node. The *-gen.c file will have, at the type
init function, a call to actually init (once) the log domain, so we
don't have to do that at node opening time anymore.

Exceptions had to be made for modules that shared a domain via an
external log domain variable.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>