Fix exiting of forked process by way of _exit()
Only one of the two sides of a fork() can exit via regular
exit(). That's because it will try to flush any open FILE* and run
atexit() functions, among other things (e.g., running C++
destructors). If both sides do it, weird things happen, which are very,
VERY hard to debug.

Since the parent is not exiting with _exit() in any of those cases, we
conclude the child process must.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>