fbp-generator: add support for DECLARE statements
Now the generator handles the declaration and usage of new types in
the FBP itself.

e.g.
    DECLARE=MyType:fbp:my-type.fbp
    my_node(MyType) OUT -> ...

In the generated C code now, we first build the node spec with a
placeholder and after we change this to the proper value, since
we need to create this new types in runtime.

Signed-off-by: Ricardo de Almeida Gonzaga <ricardo.gonzaga@intel.com>