pthread: fix building with or without pthread support
When some dependency checks were removed the pthread one went from
PTHREAD_H to PTHREAD and thus breaking the build.

However, the bigger problem is that we were always compiling with
pthread support because the dependency solver always defines
HAVE_PTHREAD_H even when we mark the config option PTHREAD as 'n'.
Previously autotools defined HAVE_PTHREAD_H when we had pthread.h
_and_ wanted to compile with pthreads support. Now that's not the
case anymore with Kbuild so we should test for PTHREAD in the code
and not HAVE_PTHREAD_H.

Note that the ldflags still need to be properly sorted out even though
everything compiles and works now.

Signed-off-by: Ulisses Furquim <ulisses.furquim@intel.com>