build: dep resolver, split rules and check imp + ldflags
This patch introduces a new implementation for dependency-resolver.py.
After a talk with Caio and Lucas De Marchi we realized we wanted to
split the rules and the checks implementation as well as reusing the
headers and C fragment based tests.

This patch also introduces the obj-*-[y.m]-extra-ldflags so we can
set the proper ldflags for each module.

PS: these two "features" needed to be kept in the same patch cause
the first one breaks the build.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>