build: get rid of USE_* configs
We don't need these USE_* configs any more, for that we need a historical
context. At first we were applying all the "extra cflags" to every module,
with that we should know the dependencies we needed - or were in use -
so we could avoid using it - i.e linking against dep libs - whenever not
a single module was using it.

After a while we understood we should not link all .so files against all
dependencies, for that we introduces the -extra-cflags variables so we
could apply the cflags only wherever required, once it's being applied
only to the actual users we don't need to declare who is using it with
the USE_* configs - but simply indicate their extra-cflags/extra-ldflags.

Having said that, this patch gets rid of all USE_* configs.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>