build: lib shared should never be disabled
The lib-shared (or src/shared) should never be disabled and should
never be builtin. This is an special/different type of module in the
build system. Now if we disable modules - since INTERNAL_SHARED is
a tristate component - it'll be set to y and the build system will
try to built it as a builtin module and will fail.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>