coap: add functions to make it possible to implement oic without coap.h
We just want to expose sol-coap.h, and exposing sol_coap_packet
would lead us to a lot of checks for the struct version
(a lot of functions use packet as args).

With these changes an user should be able to use sol_coap
to implement her own implemenation of oic or other protocols
based on that.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>