fbp-generator: write to a file only if the generation suceed
Before this patch, we were appending to the file piece by piece,
at the risk of generating the file by half if the generation fails
at some part.

Now we're creating an unnamed temporary file and when the generation
ends, we link this unnamed temporary file to a filepath that the
user provide us.

Signed-off-by: Ricardo de Almeida Gonzaga <ricardo.gonzaga@intel.com>