fbp-generator: write to a temporary file until the generation ends
Before this patch, we were appending to the output file piece by
piece, at the risk of generating the output file by half if the
generation fails at some part.

Now we're creating a temporary file and when the generation ends,
we rename to a filepath that the user provide us.

Signed-off-by: Ricardo de Almeida Gonzaga <ricardo.gonzaga@intel.com>