buil: templates
This patch introduces an small and trivial template processor to be
used on the generation of source code based on .in files - similarly
to what autotools does.

The main purpose here is to move all the logic involved on the template
processing from the core build system to the template itself.

A next step is to write a generic .in rule (i.e %.in: ...).

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>