check: allow disabling colors
Adds --color option that takes auto, always and never. If nothing is
provided, --color=auto is assumed, and will make a choice based on
whether the standard output is a tty or not.

In practice the automatic behavior will handle most cases, the options
exists to make sure we can force colors or no colors if the "smartness"
of our automatic check is not enough.

The options were modelled after git-diff.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>