platform-detect: Sanitize platform name
When platform name is fetched via environment or detected via .json
file, we need to enforce naming rules to avoid the use of invalid data
(mainly in Pin Multiplexer that will use the name to open the correct
Mux). If the platform name was provided via configuration we will trust
the information.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>