flow-packet: expose member descriptions in type structure.
The newly introduced 'data_type' allows describing the packet payload,
useful to generate code for well-known types and debug.

Also introduced 'members', and array describing each member of the
payload memory in terms of basic C types. This way we can
automatically generate converters to different languages (ie:
JavaScript or Python) as well as print out in console (in the
following commit).

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com