flow: reorganize data inside the builder
Split the data inside the builder between temporary data used during the
process and the 'type_data' that is used by the node type once it is
built.

Two major changes on how things work:

- When possible, use the original spec structures even during the
  construction process. Instead of having builder_node_spec, create a
  side structure and construct the node spec. Instead of having a
  ptr_vector of connections to later create an array -- and simply sort
  it later instead of keep sorted.

  This allow us to simply "take all" the structures being constructed
  and use in the spec, avoiding the extra allocation.

- Try to do the allocations as soon as possible, to avoid. This is
  reflected by the fact that functions for get_node_type() is not
  organized by nodes, conns, ports, but by "check validity", "allocate
  whatever needed", "fill things in".

Behavior of the public API should not change. This prepares ground for
later making the created type independent of the builder, e.g. not
depending on builder to remain alive.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>