tools: teach run-fbp-tests about TEST-EXPECTS-ERROR
Instead of depending on the existence of either -expected.txt or
-expected-fail.txt, run-fbp-tests check in the FBP file whether there's
a comment indicating that the test should fail. All expected output now
goes to -expected.txt.

The comment must be

    # TEST-EXPECTS-ERROR

used the name error here to avoid confusion, since we expect an error,
so if there's no error, then it is a failure.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>