flow: introduce simplectype type creator.
This allows creating a sol_flow_node_type from a single C function
that will be called with an event parameter for all possible events in
a type, such as opening, closing, connecting and disconnecting of
ports and processing of input ports.

The goal is to make a simple API, since the user can already create a
type in C using the structure.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com