tools: put the expected result inside the FBP test
Instead of looking for an -expected.txt file, the test runner now will
look for a comment "## TEST-OUTPUT" and consider all following the
lines (ignoring the first two columns) as part of the test output. E.g.

    ## TEST-OUTPUT
    # component-without-node.fbp:31:1 Couldn't parse statement.

will expect the output

    component-without-node.fbp:31:1 Couldn't parse statement.

The extra character gives us some freedom to include per-line
out-of-band information. E.g. indicate that the line should partially
match instead of be literally equal.

Also update the README to reflect the current state of the test tool.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>