shared: Fix resolving the root dir for repeating prefixes
It is possible to have PREFIX multiple times in the path, for example,
the user sets PREFIX to '/home/user' and soletta is being built in
'/home/user/src/soletta/build/soletta_sysroot/' the resulting rootpath
would be '/'. The solution is work from the end of the path.

The helper strrstr() is introduced for this.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>