build: split internal tests
We have 2 categories of tests, one testing the public exported API's
and those testing the "internal API's", the current only internal test
implementation is test-coap.

This patch introduces a differentiation for the mentioned categories.
Tests using the public API are now linked to libsoletta.so and uses
only objects exported there, the internal ones have access to any part
of internal symbols, so they can test any part of it.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>