i2c: Fix build when i2c-dev.h is provided by i2c-tools
The i2c-dev.h provided by i2c-tools redefines the symbols of i2c.h.
This is a workaround to avoid the build breaks in that case.

It fixes the build but some warnings are raised, the struct i2c_msg
provided by i2c-tools changes the type of the buf field from __u8* to
char*. Bringing warnings like:

 warning: pointer targets in initialization differ in signedness
 [-Wpointer-sign]

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>