Generate subflow code using sol-fbp-generator
This commit is the first step to achieve automatic code
generation for modules that depend on subflows. Calamari's
7-segments type is an abstraction for four GPIO nodes and
a controller (calamari/segments-ctl). Instead of specifying
this subflow by hand using the C api, we can now rely on
sol-fbp-generator to do the dirty work for us, by merely
writing an fbp file describing the subflow.

As soon as the build system starts calling the run-fbp-generator,
calamari-7seg.c will be gone and a calamari-7seg-gen.c will take
its place. Also, future patches for sol-fbp-generator must provide
a way to set the 'child_opts_set' callback. To overcome this
limitation, calamari-7seg.fbp enforces the GPIO node options,
which means that conffiles won't work and samples for
calamari/7seg must be run on kernels >= 3.18 to work properly.

Signed-off-by: Luiz Ywata <luizg.ywata@intel.com>