comms: Add HTTP client infrastructure
The backend is based off of the cURL library, supporting both HTTP and
HTTPS (other protocols are disabled, even if supported by the library).
Main loop integration works with both GLib and POSIX main loops, and
does not require any threading.

The API has been designed in such a way that's possible to implement it
using other backends if necessary.

This client is mostly for performing remote API calls rather than work
as a full-fledged HTTP client, since response is buffered in memory.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>