Runner: Simplify initialization
By adding extra dereference in packet_type is possible to turn
'input_nodes' and 'output_nodes' back to 'static const' structures,
simplifying the initialization and eliminating the need for asserts.

The main reason for this patch was to remove the 'static_assert'
calls because it's not supported by all target platforms (i.e. Galileo).

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>