flow: Remove GHashTable usage from resolver-conffile
This is one of the last usages of glib in the core library. Removing
this hash table usage should have little effect, as the number of
entries is the number of different external node types used.

It should be noted that atexit() is still used, and it should have no
adverse effects even if it the callback is called multiple times.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>