I2C: Remove Linux implementation details from documentation
This is a I2C abstraction not a SMBus so user don't need to know if
a SMBus or a I2C-plain transaction will be executed.
If the I2C Linux device driver that user is using don't support I2C-plain
transaction he will get warning from I2C Linux implementation.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>