flow: Add location (as in latitude/longitude) packet type
This will allow the creation of node types that provide the physical
location.  An example is going to be provided soon and will use the
free Freegeoip service, which obtains a coarse value based on the IP
address which performed the request to their servers.

In the future, the same information could be provided by GPS node
types.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>