comms: fix leak in network shutdown
Instead of deleting each element from the vector, simply clear the
vector. For regular (non-ptrs) vectors, iteration is only needed if
extra resource allocation was done and stored in the vector.

The leak itself was caused by iterating forward and mutating the vector,
when running 'make check-valgrind' there was a situation in which the
network->links vector had two elements.

The first element was deleted and the iteration stopped (because index
would be 1, but size was reduced to 1). In those cases where manual
deletion is necessary, SOL_VECTOR_FOREACH_REVERSE_IDX() should be used.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>