flow: add named options to replace array of strings usage
Use a specific type instead of "char **" (referred as strv) to pass
options that were not turned into the final options struct. It already
enforces the type to be defined for each option, so it contains more
information than before.

Instead of

    { "interval=1000", "something=\"abc\"", NULL }

a named options contains (in pseudo code) the following information

    { .members =
      { { .name = "interval", type = IRANGE, .irange = 1000 },
        { .name = "something", type = STRING, .string = "abc" },
      },
      .count = 2
    }

Both the FBP code and the conffile (now converted to JSON) must do more
work to get the options in the first format and lose information along
the way.

One consequence is that part of the job that was done by new_from_strv()
now must be done from outside. The rationale is that client-code might
have more information (e.g. JSON could accept numeric syntax for irange
value-only options if desired).

Most of the code that used strv to use named options. The conffile
remains like before, because it would only benefit from such a change
when the syntax is changed to allow those typed values. For now, it does
the conversion necessary from "strv" to the named options format.

This patch adds an enum to identify the options -- this reflects the
idea that we don't expect the types used in configuration to be
extensible -- following the previous change done for custom. A later
patch will change other parts of the code to take advantage of it.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>