flow: remove the new_options/del_options callbacks from node type
Exchange them for options_size and a pointer to the default options
data. The new_options callback was being used in two situations

- Get the default options for the node when creating a new one, this can
  be done using options_size and the default options exposed;

- Copy a child node options in static flow so it could be later modified
  and passed to the user. With options size we just make the copy
  directly, with the benefit of reusing a buffer instead of always
  allocating.

In the second case note that the strings inside the options are not
copied. The pointers remain valid, and the user of child_opt_set is not
expected to neither delete the whole options object, or make changes to
the existing strings, but provide their own new strings if needed. If
this proves to be troublesome, we can later make function signature
clearer about this.

For static flow, the default options is constructed as calls to export
options are made -- by copying the values from the child options being
exported. Like before, if the child has explicitly options set, this
will be used as parent's default.

One detail is that options_size must be at least sizeof(struct
sol_flow_node_options), since for objects that ignore options, we still
pass the empty option around (which contains sub_api == 0). We always
pass *some* option so that the code for each node type in open don't
need to perform NULL checks.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>