platform: change node type to follow the module/type pattern
We've recently changed sol-fbp-generator to infeer the node-types
headers based in the pattern module/type, so int/max would result in the
int-gen.h include, constant/boolean result in the constant-gen.h include
and so on.

The platform descriptor isn't following this pattern, platform-service
will not result in a platform-gen.h but platform-service-gen.h
instead. This patch adjusts the platform descriptor to follow the
overmentioned pattern.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>