suite.py: use a task queue
Instead of creating one thread per test, create a limited number of
threads (currently defined as os.cpu_count()) and use a task_queue for
those thread to pick up work. This scales better as we add more tests.

Task class was created to hold the input and to let the thread store the
output, replacing the 'multiprocessing' data structures.

Also, return failure exit code if one of the tests fails -- allowing
make and others to identify when tests failed in a programatic way.

Other small fixes and cosmetics to the code were made.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>