sol-util: fix size_t version and add uint64 mul/add.
Use the correct CPP define so the code compile. Add errno.h, it was
missing and thus EOVERFLOW.

Do not specify the builtin to use forcing a size, this is bad for
types that may change size depending on platform or configuration,
such as size_t. Instead use the generic version and let the compile
determine the correct one to use.

Add addition and multiplication for uint64. Not adding other types
since we can always use unsigned 64bits and check 32 bits or signed --
and that's what I'll use in the next commit introducing JSON number
parsing.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com