json: add number parsing.
JSON numbers may be represented in a variety of formats and by default
it's floating point.

As pointed out by acidx, parsing a floating point is a pain
(http://www.netlib.org/fp/dtoa.c), particularly when the precision is
large enough to have naive math to accumulate errors. Thus we take a
short path of strndupa() + strtod_l() with the newly introduced
sol_util_strtodn() with use_locale=false.

However most of our use-cases are to parse signed or unsigned integers
of some precision (32 or 64) and for those we can do a simple
conversion and avoid growing the stack or complexity of strtod().

Last but not least, we want to report errors but we want to help the
user by always returning a sane value.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com